PLAT3-4211; Removed document specific behaviors. Updated existing classes with override model parameters.
