--- base_asset/behavior/document/document_asset_filetype_rest.py
@@ -1,21 +1,0 @@
-from ...models import DocumentAssetDocumentCategoryModel
-from ..asset_filetype_rest import AssetFileTypeRestBehavior
-from .document_filetype_category_rest import DocumentFileTypeCategoryRestBehavior
-from lumavate_exceptions import ValidationException
-
-class DocumentAssetFileTypeRestBehavior(AssetFileTypeRestBehavior):
-  def __init__(self, model_class=DocumentAssetDocumentCategoryModel, data=None, category_type='filetype'):
-    super().__init__(model_class, data, category_type)
-
-  def set_asset_filetype(self, asset_id, filetype):
-    if not filetype:
-      raise ValidationException('No file type specified')
-    
-    self.delete_by_asset(asset_id)
-
-    tag = DocumentFileTypeCategoryRestBehavior().get_tag_by_name(filetype)
-    if not tag:
-      raise ValidationException('File type not found')
-
-    self.create_asset_category(asset_id, tag.id)
-
