--- base_asset/behavior/document/document_asset_tag_rest.py
@@ -1,6 +1,0 @@
-from ...models import DocumentAssetDocumentCategoryModel
-from ..asset_tag_rest import AssetTagRestBehavior
-
-class DocumentAssetTagRestBehavior(AssetTagRestBehavior):
-  def __init__(self, model_class=DocumentAssetDocumentCategoryModel, data=None):
-    super().__init__(model_class, data, 'tag')
